Change i to int, to match statusarray->nelts and silence MSVC warning.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@837884 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/client/status.c
@@ -63,7 +63,7 @@
 void
 svn_cl__print_status (apr_hash_t *statushash, apr_pool_t *pool)
 {
-  size_t i;
+  int i;
   char statuschar;
   apr_array_header_t *statusarray;
